separate the two Zendesk pipelines sharing this repository.
Currently a merge to this repository will trigger both pipelines. That is undesirable for the deduplicator because there are manual followup activities carried out by the support team after it runs. These changes will separate the 2 tasks such that only a merge to the respective code will trigger the pipeline. Updated README to explain.